+++ nevergrad/functions/functionlib.py
@@ -51,7 +51,7 @@ class ArtificialVariable:
         """
         # use random indices for blocks
         indices = self.random_state.choice(
-            self._dimension, self.block_dimension * self.num_blocks, replace=False
+            self._dimension, self.block_dimension * self.num_blocks, replace=False  # type: ignore
         ).tolist()
         indices.sort()  # keep the indices sorted sorted so that blocks do not overlap
         # Caution this is also important for split, so that splitted arrays end un in the same block
@@ -251,7 +251,7 @@ class ArtificialFunction(ExperimentFunction):
         try:
             val = float(self._aggregator(results))
             if self.zero_pen:
-                val += 1e3 * max(
+                val += 1e3 * max(  # type: ignore
                     self.translation_factor / (1e-7 + self.translation_factor + np.linalg.norm(x.flatten()))
                     - 0.75,
                     0.0,