+++ nevergrad/optimization/test_optimizerlib.py
@@ -183,6 +183,7 @@ def test_infnan(name: str) -> None:
                 "EDA",
                 "EMNA",
                 "Stupid",
+                "NEWUOA",
                 "Large",
                 "Fmin2",
                 "NLOPT",
@@ -256,6 +257,102 @@ def test_optimizers(name: str) -> None:
         check_optimizer(optimizer_cls, budget=budget, verify_value=verify)
 
 
+@pytest.mark.parametrize("name", registry)  # type: ignore
+def test_optimizers_minimal(name: str) -> None:
+    optimizer_cls = registry[name]
+    if optimizer_cls.one_shot or name in ["CM", "NLOPT_LN_PRAXIS", "ES", "RecMixES", "RecMutDE", "RecES"]:
+        return
+    if any(
+        x in str(optimizer_cls)
+        for x in [
+            "BO",
+            "Meta",
+            "Voronoi",
+            "tuning",
+            "ECMA",
+            "CMAstd",
+            "_COBYLA",
+            "Chain",
+            "CMAbounded",
+            "Tiny",
+            "para",
+            "SPSA",
+            "EDA",
+            "FCMA",
+            "Noisy",
+            "HS",
+            "CMandAS",
+            "GA",
+            "EMNA",
+            "RL",
+            "Milli",
+            "Micro",
+            "Naive",
+            "Portfo",
+            "ESCH",
+            "Multi",
+            "NGO",
+            "Discrete",
+            "MixDet",
+            "Rotated",
+            "Iso",
+            "Bandit",
+            "TBPSA",
+            "VLP",
+            "LPC",
+            "Choice",
+        ]
+    ):
+        return
+
+    def f(x):
+        return sum((x - 1.1) ** 2)
+
+    def mf(x):
+        return sum((x + 1.1) ** 2)
+
+    def f1(x):
+        return (x - 1.1) ** 2
+
+    def f2(x):
+        return (x + 1.1) ** 2
+
+    def f1p(x):
+        return sum((x - 1.1) ** 2)
+
+    def f2p(x):
+        return sum((x + 1.1) ** 2)
+
+    if "Cma" in name or "CMA" in name or "BIPOP" in name:  # Sometimes CMA does not work in dim 1 :-(
+        budget = 2000
+        if any(x in name for x in ["Large", "Tiny", "Para"]):
+            return
+        val = optimizer_cls(2, budget).minimize(f).value[0]
+        assert 1.04 < val < 1.16, f"pb with {optimizer_cls} for 1.1: {val}"
+        val = optimizer_cls(2, budget).minimize(mf).value[0]
+        assert -1.16 < val < -1.04, f"pb with {optimizer_cls} for -1.1.: {val}"
+        v = ng.p.Array(shape=(2,), upper=1.0, lower=0.0)
+        val = optimizer_cls(v, budget).minimize(f1p).value[0]
+        assert 0.94 < val < 1.06, f"pb with {optimizer_cls} for 1.: {val}"
+        v = ng.p.Array(shape=(2,), upper=0.3, lower=-0.3)
+        val = optimizer_cls(v, budget).minimize(f2p).value[0]
+        assert -0.36 < val < -0.24, f"pb with {optimizer_cls} for -0.3: {val}"
+    else:
+        budget = 500
+        if any(x in name for x in ["QO", "SODE"]):
+            return
+        val = optimizer_cls(1, budget).minimize(f).value
+        assert 1.04 < val < 1.16, f"pb with {optimizer_cls} for 1.1: {val}"
+        val = optimizer_cls(1, budget).minimize(mf).value
+        assert -1.16 < val < -1.04, f"pb with {optimizer_cls} for -1.1.: {val}"
+        v = ng.p.Scalar(upper=1.0, lower=0.0)  # type: ignore
+        val = optimizer_cls(v, budget).minimize(f1).value
+        assert 0.94 < val < 1.06, f"pb with {optimizer_cls} for 1.: {val}"
+        v = ng.p.Scalar(upper=0.3, lower=-0.3)  # type: ignore
+        val = optimizer_cls(v, budget).minimize(f2).value
+        assert -0.36 < val < -0.24, f"pb with {optimizer_cls} for -0.3: {val}"
+
+
 class RecommendationKeeper:
     def __init__(self, filepath: Path) -> None:
         self.filepath = filepath
@@ -284,12 +381,12 @@ def recomkeeper() -> tp.Generator[RecommendationKeeper, None, None]:
 def test_optimizers_recommendation(name: str, recomkeeper: RecommendationKeeper) -> None:
     if name in UNSEEDABLE:
         raise SkipTest("Not playing nicely with the tests (unseedable)")
-    if "BO" in name:
+    if "BO" in name or "EDA" in name:
         raise SkipTest("BO differs from one computer to another")
     if "SMAC" in name:
         raise SkipTest("SMAC is too slow for the 20s limit")
     if len(name) > 8:
-        raise SkipTest("BO differs from one computer to another")
+        raise SkipTest("Let us check only compact methods.")
     # set up environment
     optimizer_cls = registry[name]
     np.random.seed(None)
@@ -314,7 +411,11 @@ def test_optimizers_recommendation(name: str, recomkeeper: RecommendationKeeper)
         # with patch("bayes_opt.bayesian_optimization.acq_max", patched):
         recom = optim.minimize(fitness)
     if name not in recomkeeper.recommendations.index:
-        recomkeeper.recommendations.loc[name, :dimension] = tuple(recom.value)
+        # recomkeeper.recommendations.loc[name, :dimension] = tuple(recom.value)
+        for i in range(len(recom.value)):
+            recomkeeper.recommendations.loc[name, f"v{i}"] = recom.value[i]
+        # cummax.loc[:i, 'atr_Lts'] =
+        # cummax.iloc[:i].loc[:, 'atr_Lts'] =
         raise ValueError(
             f'Recorded the value {tuple(recom.value)} for optimizer "{name}", please rerun this test locally.'
         )
@@ -322,7 +423,7 @@ def test_optimizers_recommendation(name: str, recomkeeper: RecommendationKeeper)
     decimal = 2 if isinstance(optimizer_cls, optlib.ParametrizedBO) or "BO" in name else 5
     np.testing.assert_array_almost_equal(
         recom.value,
-        recomkeeper.recommendations.loc[name, :][:dimension],
+        np.array(recomkeeper.recommendations.loc[name, :][:dimension], float),
         decimal=decimal,
         err_msg="Something has changed, if this is normal, delete the following "
         f"file and rerun to update the values:\n{recomkeeper.filepath}",
@@ -351,7 +452,7 @@ def test_differential_evolution_popsize(name: str, dimension: int, num_workers: 
 def test_portfolio_budget() -> None:
     for k in range(3, 13):
         optimizer = optlib.Portfolio(parametrization=2, budget=k)
-        np.testing.assert_equal(optimizer.budget, sum(o.budget for o in optimizer.optims))
+        np.testing.assert_equal(optimizer.budget, sum(o.budget for o in optimizer.optims))  # type: ignore
 
 
 def test_optimizer_families_repr() -> None:
@@ -971,7 +1072,7 @@ def test_smoother() -> None:
 
 
 @pytest.mark.parametrize("n", [5, 10, 15, 25, 40])  # type: ignore
-@pytest.mark.parametrize("b_per_dim", [1, 10, 20])  # type: ignore
+@pytest.mark.parametrize("b_per_dim", [10, 20])  # type: ignore
 def test_voronoide(n, b_per_dim) -> None:
     if n < 25 or b_per_dim < 1 and not os.environ.get("CIRCLECI", False):  # Outside CircleCI, only the big.
         raise SkipTest("Only big things outside CircleCI.")
@@ -1047,7 +1148,7 @@ def test_weighted_moo_de() -> None:
         w[index] = 30.0
         DE.set_objective_weights(w)  # type: ignore
         targ = [np.array([np.cos(2 * np.pi * i / N), np.sin(2 * np.pi * i / N)]) for i in range(N)]
-        DE.minimize(lambda x: [np.linalg.norm(x - xi) for xi in targ])
+        DE.minimize(lambda x: [np.linalg.norm(x - xi) for xi in targ])  # type: ignore
         x = np.zeros(N)
         for u in DE.pareto_front():
             x = x + u.losses